split tests into multiple files, add hex colors
