+++ tests/test_multiformat.py
@@ -1,6 +1,4 @@
 import pytest
-from io import BytesIO
-from filecmp import cmp
 from multiformat.multiformat import Document
 
 
@@ -23,467 +21,3 @@ class TestInit:
         document = Document(doc_size, doc_layout)
         assert document.w == w
         assert document.h == h
-
-
-class TestDrawing:
-    def new_document(self):
-        return Document("letter", "portrait")
-
-    def test_draw_string(self):
-        document = self.new_document()
-        document.draw_string("Test", 0, 100, "left", "OpenSans-Regular", 12,
-                             (0, 0, 0))
-        assert document._document[-1] == {
-            "type": "string",
-            "string": "Test",
-            "x": 0,
-            "y": 100,
-            "alignment": "left",
-            "font": "OpenSans-Regular",
-            "size": 12,
-            "color": (0, 0, 0)
-        }
-
-    @pytest.mark.parametrize(
-        "string, x, y, alignment, font, size, color",
-        [
-            # string
-            # x
-            ("Test", -1, 0, "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", 10000, 0, "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", "a", 0, "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", None, 0, "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", "", 0, "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            # y
-            ("Test", 0, -1, "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", 0, 10000, "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", 0, "a", "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", 0, None, "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", 0, "", "left", "OpenSans-Regular", 12, (0, 0, 0)),
-            # alignment
-            ("Test", 0, 0, "top", "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", 0, 0, -1, "OpenSans-Regular", 12, (0, 0, 0)),
-            ("Test", 0, 0, None, "OpenSans-Regular", 12, (0, 0, 0)),
-            # font
-            ("Test", 0, 0, "left", "Verdana", 12, (0, 0, 0)),
-            ("Test", 0, 0, "left", -1, 12, (0, 0, 0)),
-            ("Test", 0, 0, "left", None, 12, (0, 0, 0)),
-            # size
-            ("Test", 0, 0, "left", "OpenSans-Regular", "a", (0, 0, 0)),
-            ("Test", 0, 0, "left", "OpenSans-Regular", -1, (0, 0, 0)),
-            # color
-            ("Test", 0, 0, "left", "OpenSans-Regular", 12, ("r", "g", "b")),
-            ("Test", 0, 0, "left", "OpenSans-Regular", 12, (-1, -1, -1)),
-            ("Test", 0, 0, "left", "OpenSans-Regular", 12, (256, 256, 256)),
-        ])
-    def test_draw_string_error(self, string, x, y, alignment, font, size,
-                               color):
-        document = self.new_document()
-        with pytest.raises(RuntimeError):
-            document.draw_string(string, x, y, alignment, font, size, color)
-
-    def test_draw_rectangle(self):
-        document = self.new_document()
-        document.draw_rectangle(0, 0, 200, 200, (0, 0, 0), (0, 0, 0), 1)
-        assert document._document[-1] == {
-            "type": "rectangle",
-            "x": 0,
-            "y": 0,
-            "w": 200,
-            "h": 200,
-            "fill_color": (0, 0, 0),
-            "border_color": (0, 0, 0),
-            "border_width": 1
-        }
-
-    @pytest.mark.parametrize(
-        "x,y,w,h,fill_color,border_color,border_width",
-        [
-            # blank rectangle
-            (-1, 0, 200, 200, None, (0, 0, 0), 0),
-            # x
-            (-1, 0, 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            (5000, 0, 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            ("a", 0, 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            (None, 0, 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            ("", 0, 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            # y
-            (0, -1, 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, 5000, 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, "a", 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, None, 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, "", 200, 200, (0, 0, 0), (0, 0, 0), 1),
-            # w
-            (0, 0, -1, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, 5000, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, "a", 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, None, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, "", 200, (0, 0, 0), (0, 0, 0), 1),
-            # h
-            (0, 0, 200, -1, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, 200, 5000, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, 200, "a", (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, 200, None, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, 200, "", (0, 0, 0), (0, 0, 0), 1),
-            # fill_color
-            (0, 0, 200, 200, (-1, -1, -1), (0, 0, 0), 1),
-            (0, 0, 200, 200, (256, 256, 256), (0, 0, 0), 1),
-            (0, 0, 200, 200, ("a", "b", "c"), (0, 0, 0), 1),
-            # border_color
-            (0, 0, 200, 200, (0, 0, 0), (-1, -1, -1), 1),
-            (0, 0, 200, 200, (0, 0, 0), (256, 256, 256), 1),
-            (0, 0, 200, 200, (0, 0, 0), ("a", "b", "c"), 1),
-            # border_width
-            (0, 0, 200, 200, (0, 0, 0), (0, 0, 0), -1),
-            (0, 0, 200, 200, (0, 0, 0), (0, 0, 0), "a"),
-        ])
-    def test_draw_rectangle_error(self, x, y, w, h, fill_color, border_color,
-                                  border_width):
-        document = self.new_document()
-        with pytest.raises(RuntimeError):
-            document.draw_rectangle(x, y, w, h, fill_color, border_color,
-                                    border_width)
-
-    def test_draw_line(self):
-        document = self.new_document()
-        document.draw_line(0, 0, 100, 100, 2, (0, 0, 0))
-        assert document._document[-1] == {
-            "type": "line",
-            "x": 0,
-            "y": 0,
-            "x1": 100,
-            "y1": 100,
-            "width": 2,
-            "color": (0, 0, 0)
-        }
-
-    @pytest.mark.parametrize(
-        "x,y,x1,y1,width,color",
-        [
-            # x
-            (-1, 0, 100, 100, 2, (0, 0, 0)),
-            (10000, 0, 100, 100, 2, (0, 0, 0)),
-            ("a", 0, 100, 100, 2, (0, 0, 0)),
-            (None, 0, 100, 100, 2, (0, 0, 0)),
-            ("", 0, 100, 100, 2, (0, 0, 0)),
-            # y
-            (0, -1, 100, 100, 2, (0, 0, 0)),
-            (0, 10000, 100, 100, 2, (0, 0, 0)),
-            (0, "a", 100, 100, 2, (0, 0, 0)),
-            (0, None, 100, 100, 2, (0, 0, 0)),
-            (0, "", 100, 100, 2, (0, 0, 0)),
-            # x1
-            (0, 0, -1, 100, 2, (0, 0, 0)),
-            (0, 0, 10000, 100, 2, (0, 0, 0)),
-            (0, 0, "a", 100, 2, (0, 0, 0)),
-            (0, 0, None, 100, 2, (0, 0, 0)),
-            (0, 0, "", 100, 2, (0, 0, 0)),
-            # y1
-            (0, 0, 100, -1, 2, (0, 0, 0)),
-            (0, 0, 100, 10000, 2, (0, 0, 0)),
-            (0, 0, 100, "a", 2, (0, 0, 0)),
-            (0, 0, 100, None, 2, (0, 0, 0)),
-            (0, 0, 100, "", 2, (0, 0, 0)),
-            # width
-            (0, 0, 100, 100, "a", (0, 0, 0)),
-            (0, 0, 100, 100, -1, (0, 0, 0)),
-            # color
-            (0, 0, 100, 100, 2, ("r", "g", "b")),
-            (0, 0, 100, 100, 2, (-1, -1, -1)),
-            (0, 0, 100, 100, 2, (256, 256, 256)),
-        ])
-    def test_draw_line_error(self, x, y, x1, y1, width, color):
-        document = self.new_document()
-        with pytest.raises(RuntimeError):
-            document.draw_line(x, y, x1, y1, width, color)
-
-    def test_draw_circle(self):
-        document = self.new_document()
-        document.draw_circle(0, 0, 200, (0, 0, 0), (0, 0, 0), 1)
-        assert document._document[-1] == {
-            "type": "circle",
-            "x": 0,
-            "y": 0,
-            "radius": 200,
-            "fill_color": (0, 0, 0),
-            "border_color": (0, 0, 0),
-            "border_width": 1
-        }
-
-    @pytest.mark.parametrize(
-        "x,y,radius,fill_color,border_color,border_width",
-        [
-            # blank rectangle
-            (-1, 0, 200, None, (0, 0, 0), 0),
-            # x
-            (-1, 0, 200, (0, 0, 0), (0, 0, 0), 1),
-            (5000, 0, 200, (0, 0, 0), (0, 0, 0), 1),
-            ("a", 0, 200, (0, 0, 0), (0, 0, 0), 1),
-            (None, 0, 200, (0, 0, 0), (0, 0, 0), 1),
-            ("", 0, 200, (0, 0, 0), (0, 0, 0), 1),
-            # y
-            (0, -1, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, 5000, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, "a", 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, None, 200, (0, 0, 0), (0, 0, 0), 1),
-            (0, "", 200, (0, 0, 0), (0, 0, 0), 1),
-            # radius
-            (0, 0, -1, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, "a", (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, None, (0, 0, 0), (0, 0, 0), 1),
-            (0, 0, "", (0, 0, 0), (0, 0, 0), 1),
-            # fill_color
-            (0, 0, 200, (-1, -1, -1), (0, 0, 0), 1),
-            (0, 0, 200, (256, 256, 256), (0, 0, 0), 1),
-            (0, 0, 200, ("a", "b", "c"), (0, 0, 0), 1),
-            # border_color
-            (0, 0, 200, (0, 0, 0), (-1, -1, -1), 1),
-            (0, 0, 200, (0, 0, 0), (256, 256, 256), 1),
-            (0, 0, 200, (0, 0, 0), ("a", "b", "c"), 1),
-            # border_width
-            (0, 0, 200, (0, 0, 0), (0, 0, 0), -1),
-            (0, 0, 200, (0, 0, 0), (0, 0, 0), "a"),
-        ])
-    def test_draw_circle_error(self, x, y, radius, fill_color, border_color,
-                               border_width):
-        document = self.new_document()
-        with pytest.raises(RuntimeError):
-            document.draw_circle(x, y, radius, fill_color, border_color,
-                                 border_width)
-
-
-class TestMarkers:
-    def new_document(self):
-        return Document("letter", "portrait")
-
-    def test_insert_page_break(self):
-        document = self.new_document()
-        document.draw_rectangle(0, 0, 200, 200, (0, 0, 0), (0, 0, 0), 1)
-        document.insert_page_break()
-        assert document._document[-1] == {
-            "type": "page_break",
-        }
-
-
-class TestGenerators:
-    def new_populated_document(self, size="letter"):
-        document = Document(size, "portrait")
-        document.draw_rectangle(100, 100, 400, 400, None, (50, 99, 154), 40)
-        document.draw_rectangle(700, 100, 200, 400, (24, 28, 31), None, 0)
-        document.draw_circle(160, 800, 150, None, (138, 41, 0), 20)
-        document.draw_circle(160, 1200, 150, (172, 108, 19), None, 0)
-        document.draw_circle(160, 1600, 150, (13, 117, 75), (150, 153, 155),
-                             20)
-        document.draw_line(document.w / 2, 0, document.w / 2, document.h, 30,
-                           (224, 224, 224))
-        document.draw_string("Left Align", document.w / 2, 800, "left",
-                             "OpenSans-Bold", 100, (0, 0, 0))
-
-        document.draw_string("Center Align", document.w / 2, 1000, "middle",
-                             "OpenSans-Bold", 100, (0, 0, 0))
-
-        document.draw_string("Right Align", document.w / 2, 1200, "right",
-                             "OpenSans-Bold", 100, (0, 0, 0))
-        document.draw_string("Page 1", document.w - 50, document.h - 50,
-                             "right", "OpenSans-Regular", 100, (0, 0, 0))
-        document.insert_page_break()
-        document.draw_rectangle(100, 100, 400, 400, None, (50, 99, 154), 40)
-        document.draw_rectangle(700, 100, 200, 400, (24, 28, 31), None, 0)
-        document.draw_line(document.w / 2, 0, document.w / 2, document.h, 30,
-                           (224, 224, 224))
-        document.draw_string("Left Align", document.w / 2, 800, "left",
-                             "OpenSans-Bold", 100, (0, 0, 0))
-
-        document.draw_string("Center Align", document.w / 2, 1000, "middle",
-                             "OpenSans-Bold", 100, (0, 0, 0))
-
-        document.draw_string("Right Align", document.w / 2, 1200, "right",
-                             "OpenSans-Bold", 100, (0, 0, 0))
-        document.draw_string("Page 2", document.w - 50, document.h - 50,
-                             "right", "OpenSans-Regular", 100, (0, 0, 0))
-        return document
-
-    def test_generate_image_unsupported(self):
-        document = self.new_populated_document()
-        with pytest.raises(RuntimeError):
-            document.generate_image("image_test", "tiff", size=(500, 500))
-
-    @pytest.mark.parametrize("page", [
-        ("A"),
-        (10),
-    ])
-    def test_generate_image_invalid_page(self, page):
-        document = self.new_populated_document()
-        with pytest.raises(RuntimeError):
-            document.generate_image(
-                "image_test", "png", size=(500, 500), page=page)
-
-    @pytest.mark.parametrize("image_format,size,page", [
-        ("PNG", None, None),
-        ("JPEG", None, 1),
-        ("GIF", None, 1),
-        ("PNG", (100, 2000), 1),
-        ("JPEG", (2000, 100), 2),
-        ("GIF", (2160, 3000), 2),
-        ("PNG", (3000, 2999), 2),
-    ])
-    def test_generate_image_supported(self, image_format, size, page):
-        document = self.new_populated_document()
-        f = BytesIO()
-        document.generate_image(
-            "image_test", image_format, size=size, page=page, file_object=f)
-
-    def test_generate_pdf(self):
-        document = self.new_populated_document()
-        document.author = "Person Name"
-        document.title = "The Title"
-        document.subject = "The Subject"
-        f = BytesIO()
-        document.generate_pdf("pdf_test", file_object=f)
-        document = self.new_populated_document("a4")
-        f = BytesIO()
-        document.generate_pdf("pdf_test", file_object=f)
-
-    def test_generated_image_files(self, tmpdir):
-        document = self.new_populated_document()
-        image_path = tmpdir.join("image_generation_test")
-        image_path_1 = tmpdir.join("image_generation_test.png")
-        image_path_2 = tmpdir.join("image_generation_test_2.png")
-        document.generate_image(image_path, "PNG", size=(1000, 1000))
-        assert cmp("tests/image_generation_test_control.png", image_path_1)
-        assert cmp("tests/image_generation_test_control_2.png", image_path_2)
-
-    def test_generated_pdf_file(self, tmpdir):
-        document = self.new_populated_document()
-        pdf_path = tmpdir.join("pdf_generation_test")
-        document.generate_pdf(pdf_path)
-
-
-class TestValidators:
-    def new_document(self):
-        return Document("A4", "portrait")
-
-    @pytest.mark.parametrize("x", [
-        ("100"),
-        (200),
-    ])
-    def test_validate_x_var(self, x):
-        document = self.new_document()
-        assert document._validate_x_var(x) == int(x)
-
-    @pytest.mark.parametrize("y", [
-        ("100"),
-        (200),
-    ])
-    def test_validate_y_var(self, y):
-        document = self.new_document()
-        assert document._validate_y_var(y) == int(y)
-
-    @pytest.mark.parametrize("x,w", [
-        ("100", "400"),
-        (200, 400),
-    ])
-    def test_validate_w_var(self, x, w):
-        document = self.new_document()
-        assert document._validate_w_var(x, w) == int(w)
-
-    @pytest.mark.parametrize("y,h", [
-        ("100", "400"),
-        (200, 400),
-    ])
-    def test_validate_h_var(self, y, h):
-        document = self.new_document()
-        assert document._validate_h_var(y, h) == int(h)
-
-    @pytest.mark.parametrize("param, return_value", [
-        ("left", "left"),
-        ("right", "right"),
-        ("middle", "middle"),
-        ("Left", "left"),
-        ("Right", "right"),
-        ("Middle", "middle"),
-        ("LEFT", "left"),
-        ("RIGHT", "right"),
-        ("MIDDLE", "middle"),
-    ])
-    def test_validate_alignment(self, param, return_value):
-        document = self.new_document()
-        assert document._validate_alignment(param) == return_value
-
-    @pytest.mark.parametrize("font,return_font", [
-        ("OpenSans-Regular", "OpenSans-Regular"),
-        ("OpenSans-Bold", "OpenSans-Bold"),
-        ("opensans-regular", "OpenSans-Regular"),
-        ("opensans-bold", "OpenSans-Bold"),
-        ("OPENSANS-REGULAR", "OpenSans-Regular"),
-        ("OPENSANS-BOLD", "OpenSans-Bold"),
-    ])
-    def test_validate_font(self, font, return_font):
-        document = self.new_document()
-        assert document._validate_font(font) == return_font
-
-    @pytest.mark.parametrize("font", [
-        ("Unknown-Font"),
-        (-1),
-        (None),
-    ])
-    def test_validate_font_error(self, font):
-        document = self.new_document()
-        with pytest.raises(RuntimeError):
-            document._validate_font(font)
-
-    @pytest.mark.parametrize("size", [
-        ("10"),
-        (20),
-    ])
-    def test_validate_size(self, size):
-        document = self.new_document()
-        assert document._validate_size(size) == int(size)
-
-    @pytest.mark.parametrize("string", [
-        ("test string"),
-    ])
-    def test_validate_string(self, string):
-        document = self.new_document()
-        assert document._validate_string(string) == string
-
-    @pytest.mark.parametrize("color,output", [
-        ((10, 10, 10), (10, 10, 10)),
-        (("100", "100", "100"), (100, 100, 100)),
-        ("000", (0, 0, 0)),
-        ("#000", (0, 0, 0)),
-        ("000000", (0, 0, 0)),
-        ("#000000", (0, 0, 0)),
-        ("FFF", (255, 255, 255)),
-        ("#FFF", (255, 255, 255)),
-        ("FFFFFF", (255, 255, 255)),
-        ("#FFFFFF", (255, 255, 255)),
-        ("87CEFA", (135, 206, 250)),
-        ("8B4513", (139, 69, 19)),
-        ("708090", (112, 128, 144)),
-    ])
-    def test_validate_color(self, color, output):
-        document = self.new_document()
-        assert document._validate_color(color) == output
-
-    @pytest.mark.parametrize("color,error", [
-        (("a", 10, 10), RuntimeError),
-        (("100", "b", "100"), RuntimeError),
-        (("100", None, "100"), RuntimeError),
-        (("100", "100", -1), RuntimeError),
-        (("100", "300", "-1"), RuntimeError),
-        ("F", ValueError),
-        ("#F", ValueError),
-        ("FF", ValueError),
-        ("#FF", ValueError),
-        ("FFFF", ValueError),
-        ("#FFFF", ValueError),
-        ("FFFFF", ValueError),
-        ("#FFFFF", ValueError),
-        ("FFFFFFF", ValueError),
-        ("#FFFFFFF", ValueError),
-    ])
-    def test_validate_color_error(self, color, error):
-        document = self.new_document()
-        with pytest.raises(error):
-            document._validate_color(color)
